+++ src/common/db/alembic/sqlite_versions/76ce66b69597_upgrade_to_version_1_6_0_rc3.py
@@ -1,0 +1,29 @@
+"""Upgrade to version 1.6.0-rc3
+
+Revision ID: 76ce66b69597
+Revises: 92038d1e365e
+Create Date: 2025-01-22 15:53:33.012802
+
+"""
+
+from typing import Sequence, Union
+
+from alembic import op
+import sqlalchemy as sa
+
+
+# revision identifiers, used by Alembic.
+revision: str = "76ce66b69597"
+down_revision: Union[str, None] = "92038d1e365e"
+branch_labels: Union[str, Sequence[str], None] = None
+depends_on: Union[str, Sequence[str], None] = None
+
+
+def upgrade() -> None:
+    # Update the version in bw_metadata
+    op.execute("UPDATE bw_metadata SET version = '1.6.0-rc3' WHERE id = 1")
+
+
+def downgrade() -> None:
+    # Revert the version in bw_metadata
+    op.execute("UPDATE bw_metadata SET version = '1.6.0-rc2' WHERE id = 1")