+++ src/ui/utils/gunicorn.conf.py
@@ -15,6 +15,10 @@ for deps_path in [join(sep, "usr", "share", "bunkerweb", *paths) for paths in ((
     if deps_path not in sys_path:
         sys_path.append(deps_path)
 
+from gevent import monkey
+
+monkey.patch_all()
+
 from passlib.totp import generate_secret
 
 from logger import setup_logger  # type: ignore
@@ -23,6 +27,7 @@ from app.models.ui_database import UIDatabase
 from app.utils import USER_PASSWORD_RX, check_password, gen_password_hash, get_latest_stable_release
 
 TMP_DIR = Path(sep, "var", "tmp", "bunkerweb")
+TMP_UI_DIR = TMP_DIR.joinpath("ui")
 RUN_DIR = Path(sep, "var", "run", "bunkerweb")
 LIB_DIR = Path(sep, "var", "lib", "bunkerweb")
 
@@ -62,7 +67,7 @@ chdir = join(sep, "usr", "share", "bunkerweb", "ui")
 umask = 0x027
 pidfile = PID_FILE.as_posix()
 worker_tmp_dir = join(sep, "dev", "shm")
-tmp_upload_dir = join(sep, "var", "tmp", "bunkerweb", "ui")
+tmp_upload_dir = TMP_UI_DIR.as_posix()
 secure_scheme_headers = {}
 forwarded_allow_ips = FORWARDED_ALLOW_IPS
 pythonpath = join(sep, "usr", "share", "bunkerweb", "deps", "python") + "," + join(sep, "usr", "share", "bunkerweb", "ui")
@@ -70,7 +75,7 @@ proxy_allow_ips = "*"
 casefold_http_method = True
 workers = MAX_WORKERS
 bind = f"{LISTEN_ADDR}:{LISTEN_PORT}"
-worker_class = "gthread"
+worker_class = "gevent"
 threads = int(getenv("MAX_THREADS", MAX_WORKERS * 2))
 max_requests_jitter = min(8, MAX_WORKERS)
 graceful_timeout = 30
@@ -90,6 +95,7 @@ if DEBUG:
 
 def on_starting(server):
     TMP_DIR.mkdir(parents=True, exist_ok=True)
+    TMP_UI_DIR.mkdir(parents=True, exist_ok=True)
     RUN_DIR.mkdir(parents=True, exist_ok=True)
     LIB_DIR.mkdir(parents=True, exist_ok=True)